refactor: rename all files and classes to align with updated naming conventions
