+++ common/autoware_overlay_rviz_plugin/autoware_overlay_msgs/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.5)
-project(rviz_2d_overlay_msgs)
+project(autoware_overlay_msgs)
 
 if (NOT CMAKE_CXX_STANDARD)
     set(CMAKE_CXX_STANDARD 17)
