+++ common/autoware_overlay_rviz_plugin/autoware_overlay_msgs/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>rviz_2d_overlay_msgs</name>
+  <name>autoware_overlay_msgs</name>
   <version>1.3.0</version>
   <description>Messages describing 2D overlays for RVIZ, extracted/derived from the jsk_visualization package.</description>
   <maintainer email="team-spatzenhirn@uni-ulm.de">Team Spatzenhirn</maintainer>
