+++ common/autoware_overlay_rviz_plugin/autoware_vehicle_overlay_rviz_plugin/assets/images/wheel.png
