+++ common/autoware_overlay_rviz_plugin/autoware_vehicle_overlay_rviz_plugin/include/signal_display.hpp
@@ -39,7 +39,7 @@
 #include <mutex>
 #endif
 
-namespace awf_2d_overlay_vehicle
+namespace autoware_vehicle_overlay_rviz_plugin
 {
 class SignalDisplay : public rviz_common::Display
 {
@@ -70,7 +70,7 @@ private Q_SLOTS:
 
 private:
   std::mutex mutex_;
-  awf_2d_overlay_vehicle::OverlayObject::SharedPtr overlay_;
+  autoware_vehicle_overlay_rviz_plugin::OverlayObject::SharedPtr overlay_;
   rviz_common::properties::IntProperty * property_width_;
   rviz_common::properties::IntProperty * property_height_;
   rviz_common::properties::IntProperty * property_left_;
@@ -119,6 +119,6 @@ private:
     const autoware_perception_msgs::msg::TrafficSignalArray::ConstSharedPtr msg);
   void drawWidget(QImage & hud);
 };
-}  // namespace awf_2d_overlay_vehicle
+}  // namespace autoware_vehicle_overlay_rviz_plugin
 
 #endif  // SIGNAL_DISPLAY_HPP_
