+++ common/autoware_overlay_rviz_plugin/autoware_vehicle_overlay_rviz_plugin/plugins_description.xml
@@ -1,0 +1,5 @@
+<library path="autoware_vehicle_overlay_rviz_plugin">
+    <class name="autoware_vehicle_overlay_rviz_plugin/SignalDisplay" type="autoware_vehicle_overlay_rviz_plugin::SignalDisplay" base_class_type="rviz_common::Display">
+        <description>Signal overlay plugin for the 3D view.</description>
+    </class>
+</library>
