+++ common/autoware_overlay_rviz_plugin/autoware_vehicle_overlay_rviz_plugin/src/turn_signals_display.cpp
@@ -31,7 +31,7 @@
 #include <memory>
 #include <string>
 
-namespace awf_2d_overlay_vehicle
+namespace autoware_vehicle_overlay_rviz_plugin
 {
 
 TurnSignalsDisplay::TurnSignalsDisplay() : current_turn_signal_(0)
@@ -39,7 +39,8 @@ TurnSignalsDisplay::TurnSignalsDisplay() : current_turn_signal_(0)
   last_toggle_time_ = std::chrono::steady_clock::now();
 
   // Load the arrow image
-  std::string package_path = ament_index_cpp::get_package_share_directory("awf_2d_overlay_vehicle");
+  std::string package_path =
+    ament_index_cpp::get_package_share_directory("autoware_vehicle_overlay_rviz_plugin");
   std::string image_path = package_path + "/assets/images/arrow.png";
   arrowImage.load(image_path.c_str());
 }
@@ -119,4 +120,4 @@ QImage TurnSignalsDisplay::coloredImage(const QImage & source, const QColor & co
   return result;
 }
 
-}  // namespace awf_2d_overlay_vehicle
+}  // namespace autoware_vehicle_overlay_rviz_plugin
